feat(wallet-sdk): own the accounts domain core on SDK-owned connections (Phase 2.5)
AccountRepository, AccountService, AccountsCache + accountsQueryOptions + the realtime
change-handler factory move into @agicash/wallet-sdk — the accounts domain now runs entirely
on connections the SDK already owns (db, encryption, cashu/spark wallet init, queryClient).

- src/error.ts (moved verbatim from shared/error.ts): DomainError/ConcurrencyError/
  NotFoundError/UniqueConstraintError + getErrorMessage — core domain primitives the
  repository throws; web shim keeps the 37 consumers untouched.
- src/accounts/account-repository.ts (method bodies verbatim): constructor switches to a
  typed AccountRepositoryDeps object (destructured into the same private fields); the
  wallet-init/auth-provider imports become plain SDK module imports ('../cashu', '../spark')
  — the injection promotion previously planned is no longer needed.
- src/accounts/account-service.ts (bodies verbatim): deps object; allMintKeysetsQueryOptions
  imported from '../cashu' directly; the User param narrows to a structural
  UserDefaultAccounts type (web's User satisfies it; replaced when the user domain moves).
- src/accounts/accounts-cache.ts: AccountsCache (version-guarded upsert + spark balance
  write-guard, verbatim), accountsQueryOptions (plain query-options object; staleTime
  Infinity + the load-bearing structuralSharing preserved verbatim with its comment), and
  createAccountChangeHandlers (the two realtime closures verbatim).
- src/spark-config.ts split out of spark.ts (configureSpark/getSparkConfig/sparkDebugLog) so
  light consumers like the cache don't pull the Breez connection module into their graph;
  spark.ts re-exports it (web surface unchanged).
- web keeps the React layer: useAccountRepository/useAccountService wiring (deps objects),
  useAccountsCache, useAccountChangeHandlers (wraps the factory), and all 16 read hooks
  unchanged; account-hooks re-exports AccountsCache + accountsQueryOptions so its importers
  are untouched. The two route construction sites switch to the deps-object shape.
- NEW src/accounts/accounts-cache.test.ts locks the load-bearing behaviors: version-guard
  (older/equal/newer/append), balance write-guard (same-ref on no-op), structuralSharing
  (preserves expired accounts, new wins on collision).

typecheck, biome, build all green; 146 tests (sdk 13, web 84, utils 32, cashu 35); no direct
react/DOM in SDK src beyond the documented opensecret exception.
